functions reorganized
