+++ analyzers/nuget/analyze.go
@@ -1,7 +1,6 @@
 package nuget
 
 import (
-	"fmt"
 	"path/filepath"
 	"regexp"
 
@@ -14,7 +13,6 @@ import (
 
 func (a *Analyzer) Analyze() (graph.Deps, error) {
 	log.WithField("module", a.Module).Debug("analyzing module")
-
 	dir := Dir(a.Module)
 	if dir == "." {
 		dir = ""
@@ -42,59 +40,57 @@ func (a *Analyzer) Analyze() (graph.Deps, error) {
 	case "projectjson":
 		return dotnet.ProjectGraph(filepath.Join(dir, "project.json"))
 	default:
-
 		if exists, err := files.Exists(filepath.Join(dir, "paket.lock")); exists && err == nil {
 			return paket.DependencyGraph(filepath.Join(dir, "paket.lock"))
 		}
 
-		// Try to use the resolving method js
 		if exists, err := files.Exists(filepath.Join(dir, "obj", "project.assets.json")); exists && err == nil {
-			graph, err := dotnet.ResolveStrategy(a.Module.BuildTarget, a.Module.Dir)
-			if err == nil {
-				return graph, nil
+			depGraph, err := dotnet.ResolveStrategy(a.Module.BuildTarget, a.Module.Dir)
+			if err == nil && len(depGraph.Transitive) > 0 {
+				return depGraph, nil
 			}
-			log.Warnf("Error trying to get dependencies from project.assets.json: %s", err.Error())
+			log.Warnf("There was a problem reading dependencies from project.assets.json: %s, %+v", err.Error(), depGraph)
 		}
-		log.Warn("project.assets.json has not been found")
 
-		// Use a packageReference function xml
 		var xmlProj = regexp.MustCompile(`.*\.(cs|x|vb|db|fs)proj$`)
 		if xmlProj.MatchString(a.Module.BuildTarget) {
-			if exists, err := files.Exists(a.Module.BuildTarget); !exists || err != nil {
-				fmt.Println("file doesn't exist or error", err, exists)
+			exists, err := files.Exists(a.Module.BuildTarget)
+			if exists && err == nil {
+				graph, err := dotnet.PackageReferenceGraph(a.Module.BuildTarget)
+				if err == nil && len(graph.Transitive) > 0 {
+					return graph, nil
+				}
+				log.Warnf("There was a problem analyzing the package reference file `%s`. Error: %s. Graph: %+v", a.Module.BuildTarget, err, graph)
+			} else {
+				log.Warnf("Target `%s` is unable to be read. File exists: %b. Error: %s", a.Module.BuildTarget, exists, err)
 			}
-			graph, err := dotnet.PackageReferenceGraph(a.Module.BuildTarget)
-			if err == nil && len(graph.Transitive) > 0 {
-				return graph, nil
-			}
-			log.Warnf("Error trying to read package reference file `%s`: %+v", a.Module.BuildTarget, err)
 		}
 
 		pattern := filepath.Join(a.Module.Dir, "*nuspec")
-		if fileMatches, err := filepath.Glob(pattern); err == nil && len(fileMatches) > 0 {
+		fileMatches, err := filepath.Glob(pattern)
+		if err == nil && len(fileMatches) > 0 {
 			file := fileMatches[0]
-			if exists, err := files.Exists(file); !exists || err != nil {
-				fmt.Println("file doesn't exist or error", err, exists)
-			}
 			graph, err := dotnet.NuspecGraph(file)
-			if err == nil {
-				return graph, err
+			if err == nil && len(graph.Transitive) > 0 {
+				return graph, nil
 			}
-			log.Warnf("Error trying to read package reference file `%s`: %+v", a.Module.BuildTarget, err.Error())
+			log.Warnf("There was a problem analyzing the nuspec file `%s`. Error: %s. Graph: %+v", file, err, graph)
 		}
 
 		if exists, err := files.Exists(filepath.Join(dir, "packages.config")); exists && err == nil {
 			graph, err := dotnet.PackageConfigGraph(filepath.Join(dir, "packages.config"))
-			if err == nil {
+			if err == nil && len(graph.Transitive) > 0 {
 				return graph, err
 			}
+			log.Warnf("There was a problem analyzing packages.config. Error: %s. Graph: %+v", err, graph)
 		}
 
 		if exists, err := files.Exists(filepath.Join(dir, "project.json")); exists && err == nil {
 			graph, err := dotnet.ProjectGraph(filepath.Join(dir, "project.json"))
-			if err == nil {
+			if err == nil && len(graph.Transitive) > 0 {
 				return graph, err
 			}
+			log.Warnf("There was a problem analyzing project.json. Error: %s. Graph: %+v", err, graph)
 		}
 	}
